+++ index.d.ts
@@ -1,57 +1,38 @@
 /// <reference types="node" />
-
-// accepted encoding types for fs.readFileSync
-export type Encoding =
-  | "ascii"
-  | "utf8"
-  | "utf-8"
-  | "utf16le"
-  | "ucs2"
-  | "ucs-2"
-  | "base64"
-  | "latin1"
-  | "binary"
-  | "hex";
-
-interface ParsedOutput {
-  [name: string]: string;
+export declare type Encoding = "ascii" | "utf8" | "utf-8" | "utf16le" | "ucs2" | "ucs-2" | "base64" | "latin1" | "binary" | "hex";
+export interface ParsedENVs {
+    [name: string]: string;
 }
-
-interface cachedENVFiles {
-  [name: string]: string;
-}
-
-/**
- * Parses a string or buffer into an object of ENVs.
- *
- * @param src - string or Buffer to be parsed
- * @returns a single object with parsed ENVs as { key: value } pairs
- */
-
-export function parse(src: string | Buffer): ParsedOutput;
-
+export declare type ProcessEnv = {
+    [key: string]: string;
+};
+export declare type LoadedEnvFiles = Array<{
+    path: string;
+    contents: string;
+}>;
 export interface ConfigOptions {
-  path?: string | string[]; // path to .env file
-  encoding?: Encoding; // encoding of .env file
-  debug?: string | boolean; // turn on logging for debugging purposes
+    dir?: string;
+    path?: string | string[];
+    encoding?: Encoding;
+    debug?: string | boolean;
 }
-
 export interface ConfigOutput {
-  parsed?: ParsedOutput; // parsed ENVs as key value pairs
-  cachedENVFiles?: cachedENVFiles; // cached ENVs as key value pairs
+    parsed: ParsedENVs;
+    extracted: ParsedENVs;
+    cachedEnvFiles: LoadedEnvFiles;
 }
-
+/**
+ * Parses a string, buffer, or precached envs into an object.
+ *
+ * @param src - contents to be parsed
+ * @returns an object with keys and values based on `src`
+ */
+export declare function parse(src: string | Buffer | LoadedEnvFiles): ParsedENVs;
 /**
  * Extracts and interpolates one or multiple `.env` files into an object and assigns them to {@link https://nodejs.org/api/process.html#process_process_env | `process.env`}.
  * Example: 'KEY=value' becomes { KEY: 'value' }
  *
- * @param options - accepts: { path: string | string[], debug: boolean, encoding: | "ascii" | "utf8" | "utf-8" | "utf16le" | "ucs2" | "ucs-2" | "base64" | "latin1" | "binary"| "hex" }
- * @returns a single parsed object with parsed ENVs as { key: value } pairs and an array of cached ENVs as { key: value} pairs
+ * @param options - accepts: { dir: string, path: string | string[], debug: boolean, encoding: | "ascii" | "utf8" | "utf-8" | "utf16le" | "ucs2" | "ucs-2" | "base64" | "latin1" | "binary"| "hex" }
+ * @returns a single parsed object with parsed ENVs as { key: value } pairs, a single extracted object with extracted ENVS as { key: value } pairs, and an array of cached ENVs as { path: string, contents: string} pairs
  */
-export function config(options?: ConfigOptions): ConfigOutput;
-
-/** snackables library interface */
-export interface SnackableEnv {
-  config: typeof config;
-  parse: typeof parse;
-}
+export declare function config(options?: ConfigOptions): ConfigOutput;