+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 All notable changes to this project will be documented in this file. See [SemVer 2.0](https://semver.org/) for commit guidelines.
 
+## [0.0.26] - 2020-27-10
+
+### Changed
+
+- Config method returns cachedEnvs, all `process.env` contents, and extracted ENVs as an object
+- Prevented ENVs in `process.env` from being overwritten
+- Parse method optionally parses cached ENV files and applies them to `process.env`
+
 ## [0.0.25] - 2020-23-10
 
 ### Added